Use forked process in Runner for consistent state
The current approach taken by `Runner#execute!` has the possibility to
leave a lot of junk lying around, as evidenced by the fact that we're
removing a bunch of constant definitions after each run, etc.

The cleaner approach to take, and one that doesn't run into any issues
with required files inside of our evaluated scripts, is to fork out a
separate process for the script evaluation itself, and just dump/load
the solution(s) via an IO pipe. This means our primary Runner object's
state is always clean, as any constant or class definitions occur just
in the forked "workspace" process.