Add spec suite for 2019 challenges
The heavy reliance of 2019's challenges on the Intcode computer means
that a lot of these scripts aren't testable in a normal way -- while
many of them have examples for sub-components of the challenge, most
of them are hypotheticals and don't work without the underlying code.

I _could_ possibly decompose my scripts more to test the pieces that
have examples provided for them, but I don't think it's worth the
effort.